🚚  [Move] RoundImageView class
Separated from ContactCell.swift
